requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
httpx>=0.27.0
pydantic>=2.7
pandas>=2.2
sqlalchemy>=2.0
jinja2>=3.1
pdfplumber>=0.11
faiss-cpu>=1.8
sentence-transformers>=3.0
numpy>=1.26
scikit-learn>=1.5
rapidfuzz>=3.9
langchain>=0.2
langchain-community>=0.2
rich>=13.7
fastapi>=0.115
